[HomeFragment] move file HomeScreenBackground and PrivateHomeScreenBackground to package under rocket
