+++ app/src/main/java/org/mozilla/rocket/home/ui/PrivateHomeScreenBackground.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus.home;
+package org.mozilla.rocket.home.ui;
 
 import android.app.Activity;
 import android.content.Context;
